requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
websockets==13.1
Pillow==10.4.0
python-dotenv==1.0.0
PyYAML==6.0.2
pydantic==2.9.2
loguru==0.7.3
pytest==8.0.0
pytest-asyncio==0.23.0
